some fixes, added vec2 float32 in double64 representation, library for float2 math (vector), and managing a float2 buffer array, added common math functions/operation, needs testing

+++ gmgmp/gmgmp/gmgmp.cpp
@@ -11,12 +11,12 @@
 #include <gmp/gmp.h>
 #pragma warning(pop)
 
-GMGMP_API int _fn_gmgmp()
+GMGMP_API double _fn_gmgmp()
 {
-	return 42;
+	return 42.0;
 }
 
-GMGMP_API int _fn_gmgmp_arg(int arg)
+GMGMP_API double _fn_gmgmp_arg(double arg)
 {
 	return arg;
 }
